scripts/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
requests>=2.31.0
cssutils>=2.9.0
lxml>=4.9.0
